split up types into seperate files
x
